feat: Nav, Menu 분리 및 기본 스타일 지정
