+++ fe/src/Component/Header/HeaderLeft/Menu.Styled.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const SubMenuUl = styled.ul`
+  font-family: "Noto Sans KR";
+  font-size: 14px;
+  line-height: 24px;
+  margin-top: 16px;
+`;
+
+export default SubMenuUl;
