+++ fe/src/Component/Header/HeaderLeft/Nav.jsx
@@ -1,34 +1,12 @@
-import PropTypes from "prop-types";
+import Menu from "./Menu";
 import { StyledNav, StyledUl } from "./Nav.styled";
-import { MenuDatas } from "../../../MockData/MockData";
-
-const SubMenu = ({ subMenuDatas }) => {
-  const subMenuList = subMenuDatas.map((subMenuData) => (
-    <li key={subMenuData.id}>{subMenuData.name}</li>
-  ));
-  return <ul>{subMenuList}</ul>;
-};
-
-SubMenu.propTypes = {
-  subMenuDatas: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.number.isRequired,
-      name: PropTypes.string.isRequired,
-    })
-  ).isRequired,
-};
-
-const menus = MenuDatas.map((MenuData) => (
-  <li key={MenuData.id}>
-    {MenuData.name}
-    <SubMenu subMenuDatas={MenuData.subMenu} />
-  </li>
-));
 
 const Nav = () => {
   return (
     <StyledNav>
-      <StyledUl>{menus}</StyledUl>
+      <StyledUl>
+        <Menu />
+      </StyledUl>
     </StyledNav>
   );
 };
